correct permission judgment logic

+++ router/device.js
@@ -9,26 +9,44 @@ const Device = require('../schema/device')
 
 // get device interface
 router.post('/dev', (req, res) => {
-  // user login & permission judgment
-  if (req.session.userinfo.uauth == 'admin') {
-    if (req.body.vdept == 'all') {
-      // get all device list while admin request
-      Device.find({}, { _id: 0 })
-        .then(devlist => {
-          res.json({
-            status: 'success',
-            dev: devlist
+  // user login judgment
+  if (req.session.userinfo) {
+    // user permission judgment
+    if (req.session.userinfo.uauth == 'admin') {
+      if (req.body.vdept == 'all') {
+        // get all device list while admin request
+        Device.find({}, { _id: 0 })
+          .then(devlist => {
+            res.json({
+              status: 'success',
+              dev: devlist
+            })
           })
-        })
-        .catch(err => {
-          res.json({
-            status: 'success',
-            msg: req.body.vdept
+          .catch(err => {
+            res.json({
+              status: 'success',
+              msg: req.body.vdept
+            })
           })
-        })
-    } else {
-      // get a dept device list while admin request
-      Device.find({ vdept: req.body.vdept }, { _id: 0 })
+      } else {
+        // get a dept device list while admin request
+        Device.find({ vdept: req.body.vdept }, { _id: 0 })
+          .then(devlist => {
+            res.json({
+              status: 'success',
+              dev: devlist
+            })
+          })
+          .catch(err => {
+            res.json({
+              status: 'fail',
+              msg: req.body.vdept
+            })
+          })
+      }
+    } else if (req.session.userinfo.uauth == 'company') {
+      // get owned device list while company request
+      Device.find({ vdept: req.session.userinfo.udept }, { _id: 0 })
         .then(devlist => {
           res.json({
             status: 'success',
@@ -38,25 +56,10 @@ router.post('/dev', (req, res) => {
         .catch(err => {
           res.json({
             status: 'fail',
-            msg: req.body.vdept
+            msg: req.session.userinfo.udept
           })
         })
     }
-  } else if (req.session.userinfo.uauth == 'company') {
-    // get owned device list while company request
-    Device.find({ vdept: req.session.userinfo.udept }, { _id: 0 })
-      .then(devlist => {
-        res.json({
-          status: 'success',
-          dev: devlist
-        })
-      })
-      .catch(err => {
-        res.json({
-          status: 'fail',
-          msg: req.session.userinfo.udept
-        })
-      })
   } else {
     res.json({
       status: 'error',
@@ -94,36 +97,40 @@ router.post('/dev/add', (req, res) => {
 
 // delete device interface
 router.post('/dev/del', (req, res) => {
-  if (req.session.userinfo.uauth == 'admin') {
-    // delete device while admin request
-    Device.findOneAndDelete(req.body)
-      .then(dev => {
-        res.json({
-          status: 'success',
-          msg: req.body.vidno
+  // user login judgment
+  if (req.session.userinfo) {
+    // user permission judgment
+    if (req.session.userinfo.uauth == 'admin') {
+      // delete device while admin request
+      Device.findOneAndDelete(req.body)
+        .then(dev => {
+          res.json({
+            status: 'success',
+            msg: req.body.vidno
+          })
         })
-      })
-      .catch(err => {
-        res.json({
-          status: 'fail',
-          msg: req.body.vidno
+        .catch(err => {
+          res.json({
+            status: 'fail',
+            msg: req.body.vidno
+          })
         })
-      })
-  } else if (req.session.userinfo.uauth == 'company' || req.session.userinfo.uauth == 'user') {
-    // delete device while company or user request
-    Device.findOneAndDelete({ vidno: req.body.vidno, vdept: req.session.userinfo.udept })
-      .then(dev => {
-        res.json({
-          status: 'success',
-          msg: req.body.vidno
+    } else if (req.session.userinfo.uauth == 'company' || req.session.userinfo.uauth == 'user') {
+      // delete device while company or user request
+      Device.findOneAndDelete({ vidno: req.body.vidno, vdept: req.session.userinfo.udept })
+        .then(dev => {
+          res.json({
+            status: 'success',
+            msg: req.body.vidno
+          })
         })
-      })
-      .catch(err => {
-        res.json({
-          status: 'fail',
-          msg: req.body.vidno
+        .catch(err => {
+          res.json({
+            status: 'fail',
+            msg: req.body.vidno
+          })
         })
-      })
+    }
   } else {
     res.json({
       status: 'error',
@@ -134,6 +141,7 @@ router.post('/dev/del', (req, res) => {
 
 // modify device interface
 router.post('/dev/mod', (req, res) => {
+  // user login judgment
   if (req.session.userinfo) {
     // modify device name
     Device.findOneAndUpdate({ vidno: req.body.vidno, vdept: req.session.userinfo.udept }, { $set: { vname: req.body.vname } })
@@ -159,6 +167,7 @@ router.post('/dev/mod', (req, res) => {
 
 // update device status interface
 router.post('/dev/up', (req, res) => {
+  // user login judgment
   if (req.session.userinfo) {
     // update device status
     Device.findOneAndUpdate({ vidno: req.body.vidno, vdept: req.session.userinfo.udept }, { $set: { vstat: req.body.vstat } })
